# Start adding some check_shape decorations