--- a/trieste/models/gpflow/interface.py
+++ b/trieste/models/gpflow/interface.py
@@ -19,6 +19,7 @@ from typing import Any, Optional
 
 import gpflow
 import tensorflow as tf
+from check_shapes import inherit_check_shapes
 from gpflow.models import GPModel
 from gpflow.posteriors import BasePosterior, PrecomputeCacheType
 from typing_extensions import Protocol
@@ -93,6 +94,7 @@ class GPflowPredictor(
     def model(self) -> GPModel:
         """The underlying GPflow model."""
 
+    @inherit_check_shapes
     def predict(self, query_points: TensorType) -> tuple[TensorType, TensorType]:
         mean, cov = (self._posterior or self.model).predict_f(query_points)
         # posterior predict can return negative variance values [cf GPFlow issue #1813]
@@ -109,9 +111,11 @@ class GPflowPredictor(
             )
         return mean, cov
 
+    @inherit_check_shapes
     def sample(self, query_points: TensorType, num_samples: int) -> TensorType:
         return self.model.predict_f_samples(query_points, num_samples)
 
+    @inherit_check_shapes
     def predict_y(self, query_points: TensorType) -> tuple[TensorType, TensorType]:
         return self.model.predict_y(query_points)