--- a/trieste/models/gpflux/models.py
+++ b/trieste/models/gpflux/models.py
@@ -19,6 +19,7 @@ from typing import Any, Callable, Optional
 import dill
 import gpflow
 import tensorflow as tf
+from check_shapes import inherit_check_shapes
 from gpflow.inducing_variables import InducingPoints
 from gpflux.layers import GPLayer, LatentVariableLayer
 from gpflux.models import DeepGP
@@ -277,6 +278,7 @@ class DeepGaussianProcess(
     def model_keras(self) -> tf.keras.Model:
         return self._model_keras
 
+    @inherit_check_shapes
     def sample(self, query_points: TensorType, num_samples: int) -> TensorType:
         trajectory = self.trajectory_sampler().get_trajectory()
         expanded_query_points = tf.expand_dims(query_points, -2)  # [N, 1, D]