… tests/conftest.py
… tests/unit/acquisition/function/test_function.py
… tests/unit/models/gpflow/test_sampler.py
… tests/unit/models/gpflux/test_interface.py
… tests/unit/objectives/test_multi_objectives.py
… tests/unit/test_bayesian_optimizer.py
… tests/util/models/gpflow/models.py
… trieste/models/keras/interface.py
… trieste/models/keras/models.py
… trieste/objectives/multi_objectives.py
… trieste/objectives/single_objectives.py